requirements.txt
numpy
matplotlib
tqdm
pytorch
torchvision
pyro-ppl
visdom
scikit-learn
pandas
seaborn
jsonpatch